Change worms_per_region from u32 to f32 for fractional density control
Allows sub-1 values like 0.1 for sparse worm networks. Total worms computed
as ceil(worms_per_region * num_chunks) so 0.1 × 27 chunks = 3 worms.

+++ voxel-ffi/include/voxel_ffi.h
@@ -66,7 +66,7 @@ typedef struct {
     double   detail_persistence;
     double   warp_amplitude;
     /* Worm */
-    uint32_t worms_per_region;
+    float    worms_per_region;
     float    worm_radius_min;
     float    worm_radius_max;
     float    worm_step_length;
